Changed most coordinate units to float to: a. better match Processing, so there is less casting b. let you do partial inches. int is appropriate for Processing coords as you can't have half a pixel, but we need more granularity for inches.
As a result, shape movement and sizing should be more precise and less jerky.

Also added a "Lighthouse" shape, which is a beam that extends across the width of the Wiremap and spins around at a rate you specify.

Wiremap constructor now requires you to input the depth of the actual field - this was needed for the Lighthouse and will probably be useful in the future.



git-svn-id: https://wiremaplibrary.googlecode.com/svn/trunk@10 54107582-1a4f-11de-bcda-2bb9902f8029

+++ src/wiremap/WiremapGlowingSphere.java
@@ -4,15 +4,15 @@ public class WiremapGlowingSphere extends WiremapSphere {
     private int mCoreColor;
     private final int COLOR_STEPS = 16;
 
-    public WiremapGlowingSphere(Wiremap map, int x, int y, int z,
-            int baseColor, int radius, int coreColor) {
+    public WiremapGlowingSphere(Wiremap map, float x, float y, float z,
+            int baseColor, float radius, int coreColor) {
         super(map, x, y, z, baseColor, radius);
         setCoreColor(coreColor);
     }
 
     public void display() {
         mMap.getParent().pushMatrix();
-        final int baseRadius = mRadius;
+        final float baseRadius = mRadius;
         final int baseColor = mBaseColor;
         float deltaR = mMap.getParent().red(mCoreColor) - mMap.getParent().red(mBaseColor);
         float deltaG = mMap.getParent().green(mCoreColor) - mMap.getParent().green(mBaseColor);
